modify FastAPI example
- Remove db context_storage (there is no reason to show it in an interface tutorial)
- Add endpoint output annotation
- Run pipeline in async
